requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
idna==3.10
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.5
typing-extensions==4.12.2
urllib3==2.2.3
xmltodict==0.13.0
